RPC client: Simplify command line string-to-JSON-value conversion code By default, all command line parameters are converted into JSON string values.  There is no need to manually specify the incoming type.
A binary decision "parse as string or JSON?" is all that's necessary.

Convert to a simple class, initialized at runtime startup, which offers
a quick lookup to answer "parse as JSON?" conversion question.

Future parameter conversions need only to indicate the method name
and zero-based index of the parameter needing JSON parsing.